chore(rxBulkSelect): decompose monolithic source files
