+++ src/rxBulkSelect/rxBulkSelect.module.js
@@ -1,0 +1,23 @@
+/**
+ * @ngdoc overview
+ * @name rxBulkSelect
+ * @requires rxCheckbox
+ * @description
+ * # rxBulkSelect Component
+ *
+ * A component used to perform an action on multiple items in a table.
+ *
+ * ## Directives
+ * * {@link rxBulkSelect.directive:rxBatchActions rxBatchActions}
+ * * {@link rxBulkSelect.directive:rxBulkSelect rxBulkSelect}
+ * * {@link rxBulkSelect.directive:rxBulkSelectHeaderCheck rxBulkSelectHeaderCheck}
+ * * {@link rxBulkSelect.directive:rxBulkSelectRow rxBulkSelectRow}
+ * * {@link rxBulkSelect.directive:rxBulkSelectValidate rxBulkSelectValidate}
+ *
+ * ## Services
+ * * {@link rxBulkSelect.service:NotifyProperties NotifyProperties}
+ * * {@link rxBulkSelect.service:rxBulkSelectUtils rxBulkSelectUtils}
+ */
+angular.module('encore.ui.rxBulkSelect', [
+    'encore.ui.rxCheckbox'
+]);
