+++ src/rxBulkSelect/scripts/rxBulkSelect.exercise.js
